Update OVRPlayerController.cs
Since the camera is attached to the PlayerController, and this player controller has no physical shape (i.e. currently has no mesh collider or RigidBody associated with it, I changed GravityModifier to 0.0f to keep the camera from falling to the ground. Alternatively, it could be accomplished by creating a mesh collider to represent the user's physical body, give it gravity, but set the camera to the top of the capsule so that the capsule/camera pair falls but stops once the bottom touches the floor.