Added a script and a standard XML file
All_Walls.cs will setup the scaling of the U-wall and the height of the
ceiling. To set these, modify the scaleX and scaleZ (for scale of room
size - width and length), in the XML file config.xml. To set ceiling
scale, modify the scaleY in the XML file config.xml.